pyproject.toml
[tool.poetry]
name = "server"
version = "0.1.0"
description = ""
authors = ["Romain Retureau <romain.retureau@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^4.2.7"
django-annoying = "^0.10.6"
djangorestframework = "^3.14.0"
python-dotenv = "^1.0.0"
gunicorn = "^21.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
